app/requirements.txt
requests
Pillow
Mastodon.py
Pandas
matplotlib